+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -49,7 +49,6 @@ public class UserDao {
         );
     }
 
-
     public List<User> findAll() {
         final var sql = "select id, account, password, email from users";
