+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -47,5 +47,4 @@ public class UserHistoryDao {
             DataSourceUtils.releaseConnection(conn, dataSource);
         }
     }
-
 }
