+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,58 @@
+package com.techcourse.service;
+
+import com.interface21.dao.SqlExecutionException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.domain.User;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+    private final DataSource dataSource;
+
+    public TxUserService(final UserService userService, final DataSource dataSource) {
+        this.userService = userService;
+        this.dataSource = dataSource;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createdBy) {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+
+        try {
+            connection.setAutoCommit(false);
+
+            userService.changePassword(id, newPassword, createdBy);
+
+            connection.commit();
+        } catch (Exception e) {
+            rollbackAndThrow(connection, e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+            TransactionSynchronizationManager.unbindResource(dataSource);
+        }
+    }
+
+    private void rollbackAndThrow(final Connection connection, final Exception originalException) {
+        try {
+            connection.rollback();
+        } catch (SQLException rollbackEx) {
+            originalException.addSuppressed(rollbackEx);
+        }
+        throw new SqlExecutionException(originalException);
+    }
+}
