+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -8,9 +8,8 @@ import com.interface21.jdbc.core.JdbcTemplate;
 import javax.sql.DataSource;
 
 public class MockUserHistoryDao extends UserHistoryDao {
-
-    public MockUserHistoryDao(final DataSource dataSource) {
-        super(dataSource);
+    public MockUserHistoryDao(final JdbcTemplate jdbcTemplate) {
+        super(jdbcTemplate);
     }
 
     @Override
